Simulation working. ToDo: bookeepping
